requirements.txt
openai==1.35.3
requests==2.32.3
pydub==0.25.1
pyperclip==1.9.0
ollama==0.2.1
langchain==0.2.6
langchain-community==0.2.6
PyAudio==0.2.14
pynput==1.7.7
SpeechRecognition==3.10.4
python-dotenv==1.0.1